feat(lns-cli)!: drop `lns policy`, leaving the run's prompt and the file
A network decision is recorded two ways: answer the card a run raises, or
edit `lns-local-mixin.yaml` yourself. `lns policy allow/deny/allow-tcp/
deny-tcp/list/remove` was a third, and the CLI specification (§3.2.3) does
not carry it.

Removes crates/lns-cli/src/policy/ and its registry entry, policy_cli.feature
and its step module, and rewrites the docs that taught the verbs into
hand-editing the file. The two generic assertions the policy steps happened
to own — "the command succeeds" and "the command fails with an exit code
other than 0" — move to steps/cli.rs, which other features already use.

Closes #187
