+++ crates/lns-cli/src/command.rs
@@ -73,7 +73,6 @@ pub fn registry() -> Vec<CommandSpec> {
         crate::service::SPEC,
         crate::update::SPEC,
         crate::uninstall::SPEC,
-        crate::policy::SPEC,
         crate::connector::SPEC,
         crate::config::SPEC,
         crate::login::LOGIN_SPEC,
