+++ crates/lns-cli/src/lib.rs
@@ -9,7 +9,6 @@ pub mod log;
 pub mod login;
 pub mod output;
 pub mod platform;
-pub mod policy;
 pub mod raw_mode;
 pub mod run;
 pub mod sandbox;
