+++ crates/lns-cli/tests/behaviours/sandbox_surface.feature
@@ -15,7 +15,6 @@ Feature: the sandbox is the one noun, on a two-tier surface
     And the output contains "ps"
     And the output contains "sandbox"
     And the output contains "volume"
-    And the output contains "policy"
     And the output contains "connector"
 
   Scenario: the front page hides the sandbox-only verbs
