+++ crates/lns-cli/tests/behaviours/steps/cli.rs
@@ -107,6 +107,23 @@ fn parse_error_names_flag(world: &mut BehaviourWorld, flag: String) -> Result<()
     }
 }
 
+#[then(regex = r"^the command succeeds$")]
+fn command_succeeds(world: &mut BehaviourWorld) -> Result<(), String> {
+    match world.result.as_ref() {
+        Some(run) if run.exit_code == 0 => Ok(()),
+        Some(run) => Err(format!("exited {}:\n{}", run.exit_code, run.output)),
+        None => Err("the command did not run".to_string()),
+    }
+}
+
+#[then(regex = r"^the command fails with an exit code other than 0$")]
+fn command_fails(world: &mut BehaviourWorld) -> Result<(), String> {
+    match world.result.as_ref().map(|r| r.exit_code) {
+        Some(0) | None => Err("expected a non-zero exit code".to_string()),
+        Some(_) => Ok(()),
+    }
+}
+
 #[then("no run is started")]
 fn no_run_is_started(world: &mut BehaviourWorld) -> Result<(), String> {
     let res = world.result.as_ref().ok_or("no CLI run captured")?;
